requirements.txt
# 变电站设备分割训练环境依赖

# ============================================================================
# 核心依赖
# ============================================================================
numpy>=1.19.0              # 数值计算
opencv-python>=4.5.0       # 图像处理
matplotlib>=3.3.0          # 可视化
pillow>=8.0.0              # 图像操作
pyyaml>=5.4.0              # YAML配置文件解析

# ============================================================================
# 深度学习框架
# ============================================================================
torch>=2.0.0               # PyTorch核心库
torchvision>=0.15.0        # PyTorch视觉库
torchaudio>=2.0.0          # PyTorch音频库（可选）

# ============================================================================
# YOLO训练框架
# ============================================================================
ultralytics>=8.4.0         # Ultralytics YOLO框架（支持YOLO26等）

# ============================================================================
# 工具包
# ============================================================================
tqdm>=4.62.0               # 进度条
scipy>=1.7.0               # 科学计算
scikit-learn>=0.24.0       # 机器学习工具
pandas>=1.3.0              # 数据处理
seaborn>=0.11.0            # 数据可视化

# ============================================================================
# 训练监控（可选）
# ============================================================================
tensorboard>=2.10.0        # 训练可视化
wandb>=0.12.0              # 实验跟踪（可选）

# ============================================================================
# 数据增强（可选，Ultralytics已内置）
# ============================================================================
# albumentations>=1.3.0    # 高级数据增强
# imgaug>=0.4.0            # 图像增强

# ============================================================================
# 导出部署（可选）
# ============================================================================
# onnx>=1.12.0             # ONNX导出（Ultralytics已内置）
# onnxruntime>=1.12.0      # ONNX运行时
# coremltools>=6.0         # CoreML导出

# ============================================================================
# 开发工具（可选）
# ============================================================================
# jupyter>=1.0.0           # Jupyter Notebook
# ipython>=8.0.0           # 增强的Python交互式环境
# black>=22.0.0            # 代码格式化
# flake8>=4.0.0            # 代码检查
